requirements.txt
Django==3.2
django-crispy-forms==1.14.0
python-dotenv==0.21.0